extract rom/file operation
