+++ gb-cpu/src/address_bus.rs
@@ -1,3 +1,7 @@
+pub mod operation;
+
+pub use operation::{FileOperation, RomOperation};
+
 /// AddressBus map specific range address to specific area like ROM/RAM.
 /// This Implementation of an AddressBus will be limited to 16-bit address
 pub struct AddressBus {
@@ -144,22 +148,4 @@ impl Position {
     pub fn from_offset(addr: u16, offset: u16) -> Self {
         Self::new(addr - offset, addr)
     }
-}
-
-/// RomOperation basic trait to implement for a ROM Emulator.
-/// Rom is generally Read-only so `write` is not often used
-pub trait RomOperation {
-    /// writing to rom can be use full for MBC controller to set their own registry
-    fn write(&mut self, _v: u8, addr: Position) -> Result<(), Error> {
-        Err(Error::SegmentationFault(addr.absolute))
-    }
-
-    /// read one byte of data from rom
-    fn read(&self, addr: Position) -> Result<u8, Error>;
-}
-
-/// FileOperation basic trait to implement for a RAM Emulator or other area.
-pub trait FileOperation {
-    fn write(&mut self, v: u8, addr: Position) -> Result<(), Error>;
-    fn read(&self, addr: Position) -> Result<u8, Error>;
 }
